Atualizações para SQL SERVER

+++ model/list-loginstokill.php
@@ -45,59 +45,58 @@ if ($player == 'OCI'):
 
 
     $result= $conn->sql( basename(__FILE__), 
-                        "SELECT u.username,0 as admtrigger, decode(k.username,null,0,1) as tokill
-                        FROM dba_users u
-                        LEFT JOIN adm_logins_to_kill k
-                            ON u.username = k.username
-                        -- teste
-                        inner join adm_logins l
-                            on u.username = l.username
-                        --
-                        WHERE account_status='OPEN'
-                        MINUS
-                        (
-                        SELECT grantee as username, 0 as admtrigger, 0 as tokill
-                        FROM dba_role_privs 
-                        WHERE granted_role = 'DBA'
-                        UNION
-                        SELECT u.username, 0 as admtrigger, 0 as tokill
-                        FROM dba_users u 
-                        WHERE u.account_status='OPEN'
-                            AND EXISTS
-                                ( SELECT grantee 
-                                    FROM  dba_sys_privs 
-                                    WHERE privilege = 'ADMINISTER DATABASE TRIGGER'
-                                    AND grantee = u.username
-                                )
-                        )
-                        UNION
-                        SELECT grantee as username, 1 as admtrigger, 0 as tokill
-                        FROM dba_role_privs 
-                        WHERE granted_role = 'DBA'
-                        UNION
-                        SELECT u.username, 1 as admtrigger, 0 as tokill
-                        FROM dba_users u 
-                        WHERE u.account_status='OPEN'
-                        AND EXISTS
-                            ( SELECT grantee 
-                                FROM  dba_sys_privs 
-                                WHERE privilege = 'ADMINISTER DATABASE TRIGGER'
-                                    AND grantee = u.username
-                                )
-                        ORDER BY username"
+                        "SELECT u.username, adm.admtrigger, k.tokill, l.admlogins
+                        FROM dba_users u, 
+                            (select username, 1 tokill from adm_logins_to_kill) k,
+                            (select username, 1 admlogins from adm_logins) l,
+                            ( SELECT u1.username, 1 admtrigger, 0 tokill
+                                        FROM dba_users u1,
+                                            ( SELECT grantee username, 1 as admtrigger, 0 as tokill
+                                                FROM  dba_sys_privs 
+                                                WHERE privilege = 'ADMINISTER DATABASE TRIGGER'
+                                                UNION
+                                            SELECT distinct grantee username, 1 as admtrigger, 0 as tokill
+                                            FROM  dba_role_privs 
+                                            WHERE granted_role in (select grantee from dba_sys_privs where privilege = 'ADMINISTER DATABASE TRIGGER')                                         
+                                            ) x
+                            
+                            WHERE u1.account_status='OPEN'
+                                and u1.username = x.username     
+                                ) adm
+                    WHERE u.account_status='OPEN'
+                        and u.username = adm.username(+)
+                        and u.username = k.username(+)
+                        and u.username = l.username(+)
+                        order by 1"
                         );
 
 elseif ($player == 'SQLSRV'):
 
-    $result= $conn->sql( basename(__FILE__),                       
-                      "SELECT DISTINCT l.USERNAME, 0 as ADMTRIGGER,  iif(k.username is null,0,1) as TOKILL
-                      FROM adm_logins l
-                        LEFT JOIN adm_logins_to_kill k
-                            ON l.username = k.username"
+    $result= $conn->sql( basename(__FILE__),  
+                     "SELECT name USERNAME, 0 ADMTRIGGER, 0 TOKILL, 0 ADMLOGINS from sys.database_principals where type_desc='SQL_USER'
+                      EXCEPT
+                     (SELECT lower(username) username, 0 admtrigger, 0 tokill, 0 admlogins from adm_logins
+                       UNION
+                      SELECT lower(username) username, 0 admtrigger, 0 tokill, 0 admlogins from ADM_LOGINS_TO_KILL
+                     )
+                       UNION 
+                      SELECT lower(username) username, 0 admtrigger, 1 tokill, 0 admlogins from ADM_LOGINS_TO_KILL
+                       UNION
+                      SELECT lower(username) username, 0 admtrigger, 0 tokill, 1 admlogins from adm_logins
+                      EXCEPT
+                      SELECT lower(username) username, 0 admtrigger, 0 tokill, 1 admlogins from ADM_LOGINS_TO_KILL
+                     "
                         );
 
+
+
 endif;
-                    
+
+/*
+    tokill       : (1) O registro já está na tabela adm_logins_to_kill (ícone unlock) - (0) ainda não está (ícone lock)
+    admlogins    : (1) Existe regra, portanto pode permitir o LOCK (0) Não há regra, não permitir LOCK, ou seja, enviar para ADM_LOGINS_TO_KILL
+    admtrigger   : (1) Para banco ORACLE - Usuários com GRANT ADMINISTER DATABASE TRIGGER não podem ser cortados (0) Não tem esse grant, pode sofrer KILL SESSION
+*/
 
 foreach ($result as $key => $value) {
     
@@ -111,10 +110,15 @@ foreach ($result as $key => $value) {
         echo "<td style='text-align:center'><a href='\admlogins/lockuser/$user_name'><i class='fa fa-lock'></i></a></td>";
         $msg = 'usuário HABILITADO para KILL SESSION';        
 
-    elseif (!$result[$key]['TOKILL'] && !$result[$key]['ADMTRIGGER']):
+    elseif (!$result[$key]['TOKILL'] && !$result[$key]['ADMTRIGGER'] && $result[$key]['ADMLOGINS'] ):
         echo "<tr class='notokill'>";
         echo "<td style='text-align:center'><a href='\admlogins/lockuser/$user_name'><i class='fa fa-unlock'></i></a></td>";
         $msg = 'usuário DESABILITADO para kill session ';
+
+    elseif (!$result[$key]['ADMLOGINS'] && !$result[$key]['ADMTRIGGER']):
+        echo "<tr class='admtrigger'>";
+        echo "<td></td>";		
+        $msg = 'Não há regra definida para esse usuário';
 
     else:
         echo "<tr class='admtrigger'>";
